Work around bug in CCS PCD 5.123 compiler.
A simple check for "1 == 1" was failing. It would even print the two variables and show them both as 1, but the comparison would fail. A workaround is to change both data types to be U8s and then it seems to work.

+++ tlv_ptr.c
@@ -74,8 +74,8 @@ tlv_parse_ptr (CONST void * p_buf,
     {
         CONST uint8_t * p_read  = NULL;
         CONST uint8_t * p_end   = NULL;
-        unsigned int    type    = 0;
-        unsigned int    length  = 0;
+        uint8_t         type    = 0;
+        uint8_t         length  = 0;
 
         // To avoid parsing corrupt data, we first need to scan and find the
         // end of the TLV data and then check the CRC.
